+++ GTWServer/src/main/java/gruppo05/gtwserver/db/WordDAO.java
@@ -1,8 +1,8 @@
 package gruppo05.gtwserver.db;
 
 import gruppo05.gtwserver.model.Word;
+import gruppo05.gtwserver.model.WordId;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -15,7 +15,7 @@ import java.util.Optional;
  *
  * @author francesco-vecchione
  */
-public class WordDAO implements DAO<Word>{
+public class WordDAO implements DAO<Word, WordId>{
     
     private Word mapWord(ResultSet rs) throws SQLException {
         return new Word(
@@ -25,7 +25,9 @@ public class WordDAO implements DAO<Word>{
     }
 
     @Override
-    public Optional<Word> selectById(Word modelWithId) {
+    public Optional<Word> selectById(WordId modelId) {
+        if(modelId == null) return Optional.empty();
+        
         Optional<Word> result = Optional.empty();
         
         final String query = 
@@ -35,8 +37,8 @@ public class WordDAO implements DAO<Word>{
         
         try(Connection conn = DatabaseManager.getConnection();
                 PreparedStatement cmd = conn.prepareStatement(query)) {
-            cmd.setString(1, modelWithId.getToken());
-            cmd.setInt(2, modelWithId.getSource());
+            cmd.setString(1, modelId.getToken());
+            cmd.setInt(2, modelId.getSource());
             
             try (ResultSet rs = cmd.executeQuery()) {
                 if(rs.next()) {
@@ -75,6 +77,7 @@ public class WordDAO implements DAO<Word>{
 
     @Override
     public void insert(Word model) {
+        if(model == null) return;
         
         final String query = 
                 "INSERT INTO word (token, frequency, source) " +
@@ -93,7 +96,46 @@ public class WordDAO implements DAO<Word>{
     }
 
     @Override
+    public void insertAll(List<Word> modelList) {
+        if(modelList == null || modelList.isEmpty()) return;
+        
+        final String query = 
+                "INSERT INTO word (token, frequency, source) " +
+                "VALUES (?,?,?);";
+        
+        try (Connection conn = DatabaseManager.getConnection();
+                PreparedStatement cmd = conn.prepareStatement(query)) {
+            try {
+                // Tutto deve essere eseguito in una transazione
+                conn.setAutoCommit(false);
+                
+                for(Word model : modelList) {
+                    cmd.setString(1, model.getToken());
+                    cmd.setInt(2, model.getFrequency());
+                    cmd.setInt(3, model.getSource());
+                    // Aggiungi la query al pacchetto di comandi da eseguire
+                    cmd.addBatch();
+                } 
+                
+                cmd.executeBatch();
+                conn.commit();
+            } catch (SQLException sqle) {
+                try {
+                    conn.rollback();
+                } catch (SQLException ex) {
+                    throw new SQLException("Commit fallito - Rollback fallito", ex);
+                }
+                throw new SQLException("Commit fallito - Rollback effettuato", sqle);
+            }
+        } catch (SQLException ex) {
+            // Debug: da cambiare
+            ex.printStackTrace();
+        }     
+    }
+
+    @Override
     public void update(Word model) {
+        if(model == null) return;
         
         final String query = 
                 "UPDATE word " +
@@ -113,7 +155,8 @@ public class WordDAO implements DAO<Word>{
     }
 
     @Override
-    public void delete(Word modelWithId) {
+    public void delete(WordId modelId) {
+        if(modelId == null) return;
         
         final String query = 
                 "DELETE FROM word " +
@@ -125,8 +168,8 @@ public class WordDAO implements DAO<Word>{
             // Necessario se vogliamo far rispettare i vincoli di integrità referenziale
             cmd.execute(DatabaseManager.ENABLE_FOREIGN_KEYS);
             
-            cmd.setString(1, modelWithId.getToken());
-            cmd.setInt(2, modelWithId.getSource());
+            cmd.setString(1, modelId.getToken());
+            cmd.setInt(2, modelId.getSource());
             cmd.executeUpdate();
         } catch (SQLException ex) {
             // Debug: da cambiare
